# perf: reuse PushRequest object during Merge