--- a/pilot/pkg/model/push_context_test.go
+++ b/pilot/pkg/model/push_context_test.go
@@ -113,9 +113,13 @@ func TestMergeUpdateRequest(t *testing.T) {
 
 	for _, tt := range cases {
 		t.Run(tt.name, func(t *testing.T) {
-			got := tt.left.Merge(tt.right)
+			got := tt.left.CopyMerge(tt.right)
 			if !reflect.DeepEqual(&tt.merged, got) {
-				t.Fatalf("expected %v, got %v: %v", &tt.merged, got, cmp.Diff(&tt.merged, got))
+				t.Fatalf("expected %v, got %v", &tt.merged, got)
+			}
+			got = tt.left.Merge(tt.right)
+			if !reflect.DeepEqual(&tt.merged, got) {
+				t.Fatalf("expected %v, got %v", &tt.merged, got)
 			}
 		})
 	}
@@ -126,7 +130,7 @@ func TestConcurrentMerge(t *testing.T) {
 	reqB := &PushRequest{Reason: []TriggerReason{ServiceUpdate, ProxyUpdate}}
 	for i := 0; i < 50; i++ {
 		go func() {
-			reqA.Merge(reqB)
+			reqA.CopyMerge(reqB)
 		}()
 	}
 	if len(reqA.Reason) != 0 {