--- a/pilot/pkg/xds/pushqueue.go
+++ b/pilot/pkg/xds/pushqueue.go
@@ -58,12 +58,12 @@ func (p *PushQueue) Enqueue(con *Connection, pushRequest *model.PushRequest) {
 
 	// If its already in progress, merge the info and return
 	if request, f := p.processing[con]; f {
-		p.processing[con] = request.Merge(pushRequest)
+		p.processing[con] = request.CopyMerge(pushRequest)
 		return
 	}
 
 	if request, f := p.pending[con]; f {
-		p.pending[con] = request.Merge(pushRequest)
+		p.pending[con] = request.CopyMerge(pushRequest)
 		return
 	}