On a cluster with ~10k configs (svcs, eps, etc):
Before: `All caches have been synced up in 7.978471048s, marking server ready`
After: `All caches have been synced up in 2.101581551s, marking server ready`

Currently all usages already are safe, so mutation should not cause any issue.

This is so much faster mostly because ConfigUpdates; before we are allocating a ton of maps