@hzxuzhonghu 2021-09-15
It is amazing to see so much improvement by slice and map append. Maybe better to provide a benchmark.