@istio-testing 2021-09-15
@hzxuzhonghu: The following commands are available to trigger required jobs:
* `/test unit-tests_istio`
* `/test release-test_istio`
* `/test integ-pilot-k8s-tests_istio`
* `/test integ-cni-k8s-tests_istio`
* `/test integ-security-k8s-tests_istio`
* `/test integ-telemetry-k8s-tests_istio`
* `/test integ-telemetry-mc-k8s-tests_istio`
* `/test integ-multicluster-k8s-tests_istio`
* `/test integ-istiodremote-k8s-tests_istio`
* `/test integ-distroless-k8s-tests_istio`
* `/test integ-ipv6-k8s-tests_istio`
* `/test integ-operator-controller-tests_istio`
* `/test integ-pilot-multicluster-tests_istio`
* `/test integ-pilot-istiodremote-tests_istio`
* `/test integ-security-multicluster-tests_istio`
* `/test integ-helm-tests_istio`
* `/test analyze-tests_istio`
* `/test lint_istio`
* `/test gencheck_istio`
* `/test release-notes_istio`

The following commands are available to trigger optional jobs:
* `/test benchmark_istio`
* `/test integ-telemetry-istiodless-mc-k8s-tests_istio`
* `/test integ-pilot-istiodless-multicluster-tests_istio`
* `/test integ-security-istiodremote-tests_istio`
* `/test integ-security-istiodless-multicluster-tests_istio`
* `/test integ-assertion-k8s-tests_istio`

Use `/test all` to run the following jobs that were automatically triggered:
* `unit-tests_istio`
* `release-test_istio`
* `integ-pilot-k8s-tests_istio`
* `integ-cni-k8s-tests_istio`
* `integ-security-k8s-tests_istio`
* `integ-telemetry-k8s-tests_istio`
* `integ-telemetry-mc-k8s-tests_istio`
* `integ-multicluster-k8s-tests_istio`
* `integ-istiodremote-k8s-tests_istio`
* `integ-distroless-k8s-tests_istio`
* `integ-ipv6-k8s-tests_istio`
* `integ-operator-controller-tests_istio`
* `integ-pilot-multicluster-tests_istio`
* `integ-pilot-istiodremote-tests_istio`
* `integ-security-multicluster-tests_istio`
* `integ-helm-tests_istio`
* `analyze-tests_istio`
* `lint_istio`
* `gencheck_istio`
* `release-notes_istio`



<details>

In response to [this](https://github.com/istio/istio/pull/35193#issuecomment-919662101):

>/test ?
>


Instructions for interacting with me using PR comments are available [here](https://git.k8s.io/community/contributors/guide/pull-requests.md).  If you have questions or suggestions related to my behavior, please file an issue against the [kubernetes/test-infra](https://github.com/kubernetes/test-infra/issues/new?title=Prow%20issue:) repository.
</details>