@howardjohn 2021-09-15
@hzxuzhonghu the benchmark doesn't show as much improvement since we only have `pushesMerged := 10`. With `pushesMerged := 10000`:

```
name           old time/op    new time/op    delta
PushRequest-6     639ms ± 4%     413ms ±27%  -35.37%  (p=0.002 n=6+6)

name           old alloc/op   new alloc/op   delta
PushRequest-6     950MB ± 0%     107MB ± 0%  -88.68%  (p=0.004 n=6+5)

name           old allocs/op  new allocs/op  delta
PushRequest-6     3.08M ± 0%     3.04M ± 0%   -1.32%  (p=0.002 n=6+6)
``